+++ lib/src/features/version_check/data/version_repo_.dart
@@ -1,6 +1,5 @@
 import '../domain/version_check.dart';
 
 abstract class VersionCheckRepo {
-  late final VersionCheck versionCheck;
-  Future<void> getVersionCheck();
+  Future<VersionCheck> getVersionCheck();
 }